routes are protected

+++ packages/budget-backend/src/transaction/transaction.service.ts
@@ -4,6 +4,7 @@ import { Transaction } from './transaction.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateTransactionDto } from './DTO/create-transaction.dto';
 import { Category } from 'src/category/category.entity';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TransactionService {
@@ -12,29 +13,32 @@ export class TransactionService {
     private transactionRepository: TransactionRepository,
   ) {}
 
-  async getTransaction(): Promise<Transaction[]> {
-    return await this.transactionRepository.find();
+  async getTransaction(
+    user: User): Promise<Transaction[]> {
+    return await this.transactionRepository.find({ userId: user.id });
   }
 
-  async getTransactionsById(id: string): Promise<Transaction> {
-    return await this.transactionRepository.findOne(id);
+  async getTransactionsById(
+    id: string, user: User): Promise<Transaction> {
+    return await this.transactionRepository.findOne({where: { id, userId: user.id }});
   }
 
-  async createTransaction(createTransactionDto: CreateTransactionDto) {
-    return await this.transactionRepository.createTransaction(createTransactionDto);
+  async createTransaction(
+    createTransactionDto: CreateTransactionDto, user: User) {
+    return await this.transactionRepository.createTransaction(createTransactionDto, user);
   }
 
-  async deleteTransactionById(id: string): Promise<void> {
-    const result = await this.transactionRepository.delete(id);
+  async deleteTransactionById(id: string, user: User): Promise<void> {
+    const result = await this.transactionRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`No id with "${id}" found!`);
     }
   }
   async updateCategoryById(
-    id: string, category: Category): Promise<Transaction> {
+    id: string, category: Category, user: User): Promise<Transaction> {
 
-    const transaction =  await this.getTransactionsById(id);
+    const transaction =  await this.getTransactionsById(id, user);
 
     transaction.name = category;
     await transaction.save();
@@ -45,8 +49,9 @@ export class TransactionService {
 
   async getTransactionsByMonth(
     month: number,
+    user: User,
   ): Promise<Transaction[]> {
-    return await this.transactionRepository.getTransactionsByMonth(month);
+    return await this.transactionRepository.getTransactionsByMonth(month, user);
   }
 
 }
